pyproject.toml
[tool.poetry]
name = "fixitfelix"
version = "0.1.0"
description = ""
authors = ["Carsten Zwilling <czwilling@point-8.de>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.18.2"
nptdms = "^0.23.0"
pathlib = "^1.0.1"
tqdm = "^4.44.1"
pandas = "^1.0.3"
click = "^7.1.1"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.4.4"
coverage = "^5.0.4"
black = "^19.10b0"
pytest-stub = "^0.1.0"
mock = "^4.0.2"
mypy = "^0.770"
hypothesis = "^5.8.0"
click = "^7.1.1"

[tool.black]
line-length = 80
target-version = ["py38"]

[tool.poetry.scripts]
fixit = 'fixitfelix.cli:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
